Move helper to infer-data-type.ts
